refactor: client utils into clients.base, multiplex out of the pool key
`multiplex` is scaling policy, not a build input — keying the shared renderer
pool on it would duplicate a model's tokenizers for configs that differ only
there; the first client's value wins instead. Also restores one-line docstrings
on ModelContext and RolloutSession.client.
